Major update to 1.0.0.
NB: Changes are not compatible backwards.
This version passes var to the app.js file by means of the
webpack.config.js. In this way it is possible to configure different
environments: standalone/included in the module.
Also, it can send some info to a remote API server via AJAX and display
the response.
It is also turtle's compatible.